+++ Projects/Coffice/Sources/App/Main/Search/CafeDetail/CafeDetailMenuView.swift
@@ -10,18 +10,18 @@ import ComposableArchitecture
 import SwiftUI
 
 struct CafeDetailMenuView: View {
-  private let store: StoreOf<CafeDetail>
+  private let store: StoreOf<CafeDetailMenuReducer>
 
-  init(store: StoreOf<CafeDetail>) {
+  init(store: StoreOf<CafeDetailMenuReducer>) {
     self.store = store
   }
 
   var body: some View {
-    WithViewStore(store, observe: \.selectedSubMenuType) { viewStore in
+    WithViewStore(store) { viewStore in
       VStack(spacing: 0) {
         cafeSubMenuButtonView
 
-        switch viewStore.state {
+        switch viewStore.selectedSubMenuType {
         case .detailInfo:
           detailInfoMenuView
         case .review:
@@ -29,6 +29,132 @@ struct CafeDetailMenuView: View {
         }
       }
       .padding(.top, 16)
+      .onAppear {
+        viewStore.send(.onAppear)
+      }
+      .sheet(
+        item: viewStore.binding(\.$cafeReviewWriteState),
+        content: { viewState in
+          CafeReviewWriteView(
+            store: store.scope(
+              state: { _ in viewState },
+              action: CafeDetailMenuReducer.Action.cafeReviewWrite(action:)
+            )
+          )
+        }
+      )
+      .popup(
+        isPresented: viewStore.binding(\.$isReviewModifySheetPresented),
+        view: {
+          VStack(spacing: 12) {
+            Button {
+              viewStore.send(.reviewEditSheetButtonTapped)
+            } label: {
+              Text("수정하기")
+                .foregroundColor(CofficeAsset.Colors.grayScale9.swiftUIColor)
+                .applyCofficeFont(font: .button)
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .frame(height: 52)
+                .padding(.top, 20)
+            }
+
+            Button {
+              viewStore.send(.reviewDeleteSheetButtonTapped)
+            } label: {
+              Text("삭제하기")
+                .foregroundColor(CofficeAsset.Colors.grayScale9.swiftUIColor)
+                .applyCofficeFont(font: .button)
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .frame(height: 52)
+            }
+
+            Spacer()
+          }
+          .padding(.horizontal, 20)
+          .frame(height: 196)
+          .background(CofficeAsset.Colors.grayScale1.swiftUIColor)
+          .cornerRadius(12, corners: [.topLeft, .topRight])
+        },
+        customize: { popup in
+          popup
+            .type(.toast)
+            .position(.bottom)
+            .isOpaque(true)
+            .closeOnTapOutside(true)
+            .backgroundColor(CofficeAsset.Colors.grayScale10.swiftUIColor.opacity(0.4))
+            .dismissCallback {
+              viewStore.send(.reviewModifySheetDismissed)
+            }
+        }
+      )
+      .popup(
+        isPresented: viewStore.binding(\.$isReviewReportSheetPresented),
+        view: {
+          VStack(spacing: 12) {
+            Button {
+              viewStore.send(.reviewReportSheetButtonTapped)
+            } label: {
+              Text("신고하기")
+                .foregroundColor(CofficeAsset.Colors.grayScale9.swiftUIColor)
+                .applyCofficeFont(font: .button)
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .frame(height: 52)
+                .padding(.top, 20)
+            }
+
+            Spacer()
+          }
+          .padding(.horizontal, 20)
+          .frame(height: 132)
+          .background(CofficeAsset.Colors.grayScale1.swiftUIColor)
+          .cornerRadius(12, corners: [.topLeft, .topRight])
+        },
+        customize: { popup in
+          popup
+            .type(.toast)
+            .position(.bottom)
+            .isOpaque(true)
+            .closeOnTapOutside(true)
+            .backgroundColor(CofficeAsset.Colors.grayScale10.swiftUIColor.opacity(0.4))
+        }
+      )
+      .popup(
+        item: viewStore.binding(\.$deleteConfirmBottomSheetState),
+        itemView: { sheetState in
+          BottomSheetView(
+            store: store.scope(
+              state: { _ in sheetState },
+              action: CafeDetailMenuReducer.Action.bottomSheet
+            ),
+            bottomSheetContent: viewStore.bottomSheetType.content
+          )
+        },
+        customize: { BottomSheetContent.customize($0) }
+      )
+      .sheet(
+        item: viewStore.binding(\.$webViewState),
+        content: { viewState in
+          VStack(spacing: 0) {
+            CommonWebView(
+              store: store.scope(
+                state: { _ in viewState },
+                action: CafeDetailMenuReducer.Action.commonWebReducerAction
+              )
+            )
+          }
+          .customNavigationBar(
+            centerView: { EmptyView() },
+            leftView: { EmptyView() },
+            rightView: {
+              Button {
+                viewStore.send(.dismissWebView)
+              } label: {
+                CofficeAsset.Asset.close40px.swiftUIImage
+              }
+            }
+          )
+        }
+      )
     }
   }
 }
@@ -166,7 +292,7 @@ extension CafeDetailMenuView {
 
           VStack(alignment: .leading, spacing: 0) {
             Button {
-              viewStore.send(.toggleToPresentTextForTest)
+              viewStore.send(.toggleToPresentRunningTime)
             } label: {
               HStack(alignment: .top) {
                 Text(viewStore.cafe?.openingInformation?.quickFormattedString ?? "-")
@@ -177,7 +303,7 @@ extension CafeDetailMenuView {
                   .padding(.top, 4)
                 if let openingInfo = viewStore.cafe?.openingInformation,
                    openingInfo.is24Open.isFalse {
-                  viewStore.runningTimeDetailInfoArrowImageAsset.swiftUIImage
+                  Image(viewStore.runningTimeDetailInfoArrowImageName)
                     .padding(.top, 2)
                 }
               }
@@ -358,12 +484,12 @@ extension CafeDetailMenuView {
   }
 }
 
-struct CafeSearchDetailMenuView_Previews: PreviewProvider {
+struct CafeDetailMenuView_Previews: PreviewProvider {
   static var previews: some View {
     CafeDetailMenuView(
       store: .init(
-        initialState: .init(cafeId: 1),
-        reducer: CafeDetail()
+        initialState: .init(),
+        reducer: CafeDetailMenuReducer()
       )
     )
   }
